mod_rewrite.xml: add an example about override behavior
A user in PR 63051 was confused about the default behavior
of mod_rewrite (override) when merging two sections in the
directory context at the same nesting level. I was confused
as well, but after a quick look at the code I figured out that
RewriteOptions Inherit seems to solve the issue (naming is,
in this case, a bit counter intuitive).



git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1853417 13f79535-47bb-0310-9956-ffa450edef68