Merge r1768070, r1768120, r1768225, r1769712, r1769737, r1774610, r1828624, r1828626, r1855701 from trunk:
And the socache provider interface for Redis


And provide some nice usage stats. Could be prettier, for sure.

Needs LOGNO #s

reorganize mod_socache_redis so it compiles w/o apr-redis

It looks like this was the original intent. The provider
is not registered if apr-redis is missing.

It was previously failing to compile due to the dirconf being
hidden behind the #ifdef guard but not the routine module
code that used it.




Fix some HTML tags
</br> --> <br />

Couls also be <br> or <br/> but some (very) old clients don't like it.
httpd is not really consistent on it. (personnlaly <br> would be just fine for me but <br /> is the most commonly used form)

related to http://svn.apache.org/viewvc?view=revision&revision=1774602


Use apr_pcalloc in the 'create' function. All the fields are initialized in the 'init' function, but it is It is cleaner and more future proof.

Small style correction (remove extra spaces)

Initial doc for the sosache_redis module.

Corrections and improvements welcome.

Update compatibility notes for mod_socache_redis
Submitted by: jim, covener, jailletc36, jim, jailletc36, jailletc36, jailletc36
Reviewed by: jim, minfrin, covener


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1855817 13f79535-47bb-0310-9956-ffa450edef68